+++ src/api/api.js
@@ -1,7 +1,13 @@
 import qs from 'querystring'
 import fetch from 'node-fetch'
 
-import { shapePrices, shapeQuote, shapeNews, shapeWatchlist } from './shape.js'
+import {
+  shapePrices,
+  shapeQuote,
+  shapeNews,
+  shapeWatchlist,
+  shapeProfile,
+} from './shape.js'
 import { config } from '../util/config.js'
 const token = config.IEX_PUB_KEY
 
@@ -32,7 +38,7 @@ export async function getPrices(_ws, c) {
 
   let response = await fetch(url)
   if (!response.ok) {
-    throw response.statusText
+    throw response
   }
 
   const data = await response.json()
@@ -43,7 +49,7 @@ export async function getQuote(symbol) {
   const url = buildURL(`stock/${symbol}/quote`)
   let response = await fetch(url)
   if (!response.ok) {
-    throw response.statusText
+    throw response
   }
 
   response = await response.json()
@@ -54,7 +60,7 @@ export async function getNews(symbol) {
   const url = buildURL(`stock/${symbol}/news/last/10`)
   let response = await fetch(url)
   if (!response.ok) {
-    throw response.statusText
+    throw response
   }
 
   response = await response.json()
@@ -74,4 +80,26 @@ export async function getWatchlist(list) {
 
   response = await response.json()
   return shapeWatchlist(response)
+}
+
+export async function getProfile(symbol) {
+  let urls = [
+    buildURL(`stock/${symbol}/company`),
+    buildURL(`stock/${symbol}/stats`),
+    buildURL(`stock/${symbol}/earnings/1`, { period: 'quarter' }),
+  ]
+
+  const data = await Promise.all(
+    urls.map(async (url) => {
+      let response = await fetch(url)
+      if (!response.ok) {
+        throw response
+      }
+
+      response = await response.json()
+      return response
+    }),
+  )
+
+  return shapeProfile(data)
 }