+++ src/ui/repl.js
@@ -1,33 +1,30 @@
 import blessed from 'blessed'
 
-import { help, intro } from './help.js'
-import { update } from './update.js'
+import { evaluate } from './evaluate.js'
+import { intro } from './help.js'
 
 export function buildRepl(ws, c) {
   const repl = ws.grid.set(...c.yxhw, blessed.form, { keys: true })
 
   // console display (optional), otherwise commands just have effects and don't
   // report
-  const output = blessed.textarea({
+  const output = blessed.text({
     parent: repl,
     name: 'output',
-    height: '75%',
+    // inputs
+    keys: false,
+    mouse: true,
+    scrollable: true,
+    // display
     tags: true,
-    style: {
-      scrollbar: {
-        bg: 'blue',
-      },
-    },
+    height: '75%',
   })
   c.output = output
-  // init repl history
-  const history = []
 
   // add printLines to c
   ws.printLines = function (text) {
-    if (typeof text == 'string') text = text.split('\n')
-    history.push(...text)
-    output.setValue(history.slice(-output.height).join('\n'))
+    c.output.pushLine(text)
+    c.output.setScrollPerc(100)
   }
 
   // init welcome text
@@ -37,9 +34,11 @@ export function buildRepl(ws, c) {
   const input = blessed.textbox({
     parent: repl,
     name: 'input',
+    // inputs
+    inputOnFocus: true,
+    // styles
     bottom: 0,
     height: 3,
-    inputOnFocus: true,
     border: { type: 'line' },
     style: {
       border: { fg: 'gray' },
@@ -49,6 +48,7 @@ export function buildRepl(ws, c) {
     },
   })
 
+  input.unkey(['up', 'down'])
   // handle submit
   input.key('enter', function () {
     repl.submit()
@@ -56,13 +56,13 @@ export function buildRepl(ws, c) {
   repl.on('submit', async function (data) {
     // push last command
     ws.printLines('{bold}> {/}' + data.input)
+    // clear input and refocus
+    input.clearValue()
+    input.focus()
 
     // parse and handle input
     await evaluate(ws, data.input)
 
-    // clear input and refocus
-    input.clearValue()
-    input.focus()
     ws.screen.render()
   })
 
@@ -70,110 +70,4 @@ export function buildRepl(ws, c) {
   // add to curScreen
   ws.repl = repl
   ws.screen.render()
-}
-
-// helpers
-
-function evaluate(ws, input) {
-  const commands = {
-    undefined: update,
-    '#': chart,
-    '!': news,
-    '=': watchlist,
-    help,
-    h: help,
-    exit,
-    quit: exit,
-  }
-  const component = ws.activeComponent
-
-  let words = input.split(/\s+/g)
-
-  // define command to execute
-  let command = commands[words.find((w) => commands[w])]
-
-  // define params
-  let symbol = words.find((w) => w[0] == '$')
-  component.symbol = symbol ? symbol.substring(1) : component.symbol
-  let time = words.find((w) => /(?<=:)\S+/.test(w))
-  // set c.time & c.series
-  if (time) parseTime(ws, component, time)
-  if (command == commands[undefined] && !symbol && !time) {
-    ws.printLines(`{red-fg}err:{/} no valid command found\r`)
-    help(ws, component, ['help'])
-    return
-  }
-
-  // execute command
-  return command(ws, component, words)
-}
-
-function watchlist(ws, component) {
-  let listOptions = ws.options.components.find((c) => c.type == 'watchlist')
-  if (!listOptions) {
-    listOptions = {
-      type: 'watchlist',
-      yxhw: [0, 0, 12, 9],
-      symbol: component.symbol,
-    }
-    ws.options.components.push(listOptions)
-  }
-
-  update(ws, listOptions)
-}
-
-function chart(ws, component) {
-  let chartOptions = ws.options.components.find((c) =>
-    ['line'].includes(c.type),
-  )
-  chartOptions.symbol = component.symbol
-  if (component.time) parseTime(ws, chartOptions, component.time)
-
-  update(ws, chartOptions)
-}
-
-function news(ws, component) {
-  let newsOptions = ws.options.components.find((c) => c.type == 'news')
-  if (!newsOptions) {
-    newsOptions = {
-      type: 'news',
-      symbol: component.symbol,
-      yxhw: [0, 0, 12, 9],
-    }
-    ws.options.components.push(newsOptions)
-  } else {
-    newsOptions.symbol = component.symbol
-    newsOptions.time = component.time
-  }
-
-  update(ws, newsOptions)
-}
-
-/** time is a string, a valid number/interval combination, should not include
- * :-prefix*/
-export function parseTime(ws, c, time) {
-  // handle intraday
-  const intra = time.match(/([\d.]+)(min|h)/)
-  if (intra) {
-    c.series = 'intra'
-    c._time = { chartLast: +intra[1] * (intra[2] == 'h' ? 60 : 1) }
-  } else {
-    if (!ws.validUnits.includes(time.substring(1))) {
-      ws.printLines(`{red-fg}err:{/} invalid time`)
-      help(ws, c, ['h', ':'])
-      return
-    }
-    c.series = 'hist'
-    c._time = time.substring(1)
-  }
-  c.time = time
-}
-
-function exit(ws) {
-  setTimeout(() => {
-    console.log('exiting...')
-    ws.screen.destroy()
-    process.exit(0)
-  }, 800)
-  ws.printLines('{#abf-fg}goodbye...{/}')
 }